FSR-54: Refactoring following code analysis.
